+++ src/api/team-mood-tracker/get/queries/useViewReportResponse.ts
@@ -2,17 +2,17 @@ import { useQuery } from '@tanstack/react-query';
 
 import queryKeys from '@common/constants/query-key.constants';
 
-import { viewSurveyResponse } from '@features/team-mood-tracker/apis/get/view-survey-response';
+import { viewReportResponse } from '@/api/team-mood-tracker/get/apis/view-report-response';
 
 /**
- * @description 팀 설문 응답 데이터를 조회하는 useQuery 커스텀 훅
+ * @description 팀 분위기 리포트 데이터를 조회하는 useQuery 커스텀 훅
  * @param moodTrackerHashedId 조회할 설문의 해시된 ID
  */
 
-export const useViewSurveyResponse = (moodTrackerHashedId: string) => {
+export const useViewReportResponse = (moodTrackerHashedId: string) => {
   return useQuery({
-    ...queryKeys.moodTracker.detail(moodTrackerHashedId),
-    queryFn: () => viewSurveyResponse({ moodTrackerHashedId }),
+    ...queryKeys.moodTracker.report(moodTrackerHashedId),
+    queryFn: () => viewReportResponse({ moodTrackerHashedId }),
 
     enabled: !!moodTrackerHashedId,
     staleTime: 1000 * 60 * 5,
