+++ src/api/team-mood-tracker/get/queries/useViewSurveyResponse.ts
@@ -2,17 +2,17 @@ import { useQuery } from '@tanstack/react-query';
 
 import queryKeys from '@common/constants/query-key.constants';
 
-import { viewReportResponse } from '@features/team-mood-tracker/apis/get/view-report-response';
+import { viewSurveyResponse } from '@/api/team-mood-tracker/get/apis/view-survey-response';
 
 /**
- * @description 팀 분위기 리포트 데이터를 조회하는 useQuery 커스텀 훅
+ * @description 팀 설문 응답 데이터를 조회하는 useQuery 커스텀 훅
  * @param moodTrackerHashedId 조회할 설문의 해시된 ID
  */
 
-export const useViewReportResponse = (moodTrackerHashedId: string) => {
+export const useViewSurveyResponse = (moodTrackerHashedId: string) => {
   return useQuery({
-    ...queryKeys.moodTracker.report(moodTrackerHashedId),
-    queryFn: () => viewReportResponse({ moodTrackerHashedId }),
+    ...queryKeys.moodTracker.detail(moodTrackerHashedId),
+    queryFn: () => viewSurveyResponse({ moodTrackerHashedId }),
 
     enabled: !!moodTrackerHashedId,
     staleTime: 1000 * 60 * 5,
