--- src/features/team-mood-tracker/constants/end-point.constants.ts
@@ -1,4 +1,0 @@
-export const MOOD_TRACKER_API_ENDPOINTS = {
-  REPORTS: (moodTrackerHashedId: string) => `/mood-trackers/${moodTrackerHashedId}/reports`,
-  RESPONSES: (moodTrackerHashedId: string) => `/mood-trackers/${moodTrackerHashedId}/responses`,
-} as const;
